+++ packages/node-core/src/indexer/entities/Poi.entity.ts
@@ -10,7 +10,7 @@ export interface ProofOfIndex {
   parentHash?: Uint8Array;
   operationHashRoot: Uint8Array;
   mmrRoot?: Uint8Array;
-  projectId: string;
+  projectId?: string;
 }
 
 export interface PoiModel extends Model<ProofOfIndex>, ProofOfIndex {}
@@ -18,8 +18,8 @@ export interface PoiModel extends Model<ProofOfIndex>, ProofOfIndex {}
 export type PoiRepo = typeof Model & {
   new (values?: unknown, options?: BuildOptions): PoiModel;
 };
-
-export function PoiFactory(sequelize: Sequelize, schema: string): PoiRepo {
+// Project Id is deprecated
+export function PoiFactoryDeprecate(sequelize: Sequelize, schema: string): PoiRepo {
   return <PoiRepo>sequelize.define(
     `_poi`,
     {
@@ -55,6 +55,47 @@ export function PoiFactory(sequelize: Sequelize, schema: string): PoiRepo {
         type: DataTypes.STRING,
         allowNull: false,
         defaultValue: '0',
+      },
+    },
+    {
+      freezeTableName: true,
+      schema: schema,
+      indexes: [{fields: ['hash']}],
+    }
+  );
+}
+
+export function PoiFactory(sequelize: Sequelize, schema: string): PoiRepo {
+  return <PoiRepo>sequelize.define(
+    `_poi`,
+    {
+      id: {
+        type: DataTypes.INTEGER,
+        primaryKey: true,
+      },
+      chainBlockHash: {
+        type: DataTypes.BLOB,
+        allowNull: false,
+        unique: true,
+      },
+      hash: {
+        type: DataTypes.BLOB,
+        allowNull: false,
+        unique: true,
+      },
+      parentHash: {
+        type: DataTypes.BLOB,
+        allowNull: false,
+        unique: true,
+      },
+      operationHashRoot: {
+        type: DataTypes.BLOB,
+        allowNull: false,
+      },
+      mmrRoot: {
+        type: DataTypes.BLOB,
+        allowNull: true,
+        unique: true,
       },
     },
     {