+++ packages/node-core/src/indexer/project.service.ts
@@ -5,6 +5,8 @@ import assert from 'assert';
 import {isMainThread} from 'worker_threads';
 import {Inject} from '@nestjs/common';
 import {EventEmitter2} from '@nestjs/event-emitter';
+import {MetadataKeys} from '@subql/node-core/indexer/entities';
+import {CacheMetadataModel} from '@subql/node-core/indexer/storeCache';
 import {Sequelize} from 'sequelize';
 import {ApiService} from '../api.service';
 import {NodeConfig} from '../configure';
@@ -164,6 +166,7 @@ export abstract class BaseProjectService<API extends ApiService, DS extends {sta
       'processedBlockCount',
       'lastFinalizedVerifiedHeight',
       'schemaMigrationCount',
+      'deployments',
     ] as const;
 
     const existing = await metadata.findMany(keys);
@@ -185,8 +188,7 @@ export abstract class BaseProjectService<API extends ApiService, DS extends {sta
     } else {
       // Check if the configured genesisHash matches the currently stored genesisHash
       assert(
-        // Configured project yaml genesisHash only exists in specVersion v0.2.0, fallback to api fetched genesisHash on v0.0.1
-        (this.project.network.chainId ?? genesisHash) === existing.genesisHash,
+        genesisHash === existing.genesisHash,
         'Specified project manifest chain id / genesis hash does not match database stored genesis hash, consider cleaning project schema using --force-clean'
       );
     }
@@ -209,10 +211,11 @@ export abstract class BaseProjectService<API extends ApiService, DS extends {sta
     if (!existing.schemaMigrationCount) {
       metadata.set('schemaMigrationCount', 0);
     }
-
     if (!existing.startHeight) {
       metadata.set('startHeight', this.getStartBlockFromDataSources());
     }
+
+    await this.syncDeployments(existing, metadata);
   }
 
   protected async getMetadataBlockOffset(): Promise<number | undefined> {
@@ -221,6 +224,30 @@ export abstract class BaseProjectService<API extends ApiService, DS extends {sta
 
   protected async getLastProcessedHeight(): Promise<number | undefined> {
     return this.storeService.storeCache.metadata.find('lastProcessedHeight');
+  }
+
+  private async syncDeployments(existing: Partial<MetadataKeys>, metadata: CacheMetadataModel): Promise<void> {
+    if (!existing.deployments) {
+      const deployments: Record<number, string> = {};
+      //If metadata never record deployment, we are safe to use start height
+      deployments[existing.startHeight ?? this.getStartBlockFromDataSources()] = this.project.id;
+      metadata.set('deployments', JSON.stringify(deployments));
+    } else {
+      const deployments = JSON.parse(existing.deployments) as Record<number, string>;
+      const values = Object.values(deployments);
+      const lastDeployment = values[values.length - 1];
+      // avoid 0
+      if (lastDeployment !== this.project.id) {
+        const newDeploymentStart = await this.getLastProcessedHeight();
+        if (newDeploymentStart === undefined) {
+          throw new Error(`Try to record new deployment failed, unable to find last processed height from metadata`);
+        }
+        logger.warn(`Project deployment upgrade found, start height: ${newDeploymentStart}, ID: ${this.project.id}`);
+        deployments[newDeploymentStart] = this.project.id;
+        metadata.set('deployments', JSON.stringify(deployments));
+      }
+      // if we found this deployment, all good.
+    }
   }
 
   private async getStartHeight(): Promise<number> {