+++ packages/node-core/src/indexer/testing.service.ts
@@ -59,6 +59,7 @@ export abstract class TestingService<A, SA, B, DS> {
       const option: SandboxOption = {
         root: this.project.root,
         entry: file,
+        chainId: this.project.network.chainId,
       };
 
       return new TestSandbox(option, this.nodeConfig);
@@ -165,6 +166,7 @@ export abstract class TestingService<A, SA, B, DS> {
       let failedTests = 0;
       for (let i = 0; i < test.expectedEntities.length; i++) {
         const expectedEntity = test.expectedEntities[i];
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         const actualEntity = await store.get(expectedEntity._name!, expectedEntity.id);
         const attributes = actualEntity as unknown as CreationAttributes<Model>;
         const failedAttributes: string[] = [];