+++ packages/node-core/src/indexer/unfinalizedBlocks.service.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import assert from 'assert';
+import {isMainThread} from 'worker_threads';
 import {last} from 'lodash';
 import {NodeConfig} from '../configure';
 import {getLogger} from '../logger';
@@ -24,6 +25,7 @@ type UnfinalizedBlocks = Header[];
 export interface IUnfinalizedBlocksService<B> {
   init(reindex: (targetHeight: number) => Promise<void>): Promise<number | undefined>;
   processUnfinalizedBlocks(block: B | undefined): Promise<number | undefined>;
+  processUnfinalizedBlockHeader(header: Header | undefined): Promise<number | undefined>;
   resetUnfinalizedBlocks(): void;
   resetLastFinalizedVerifiedHeight(): void;
 }
@@ -89,9 +91,9 @@ export abstract class BaseUnfinalizedBlocksService<B> implements IUnfinalizedBlo
     return this.finalizedHeader.blockHeight;
   }
 
-  async processUnfinalizedBlocks(block?: B): Promise<number | undefined> {
-    if (block) {
-      this.registerUnfinalizedBlock(this.blockToHeader(block));
+  async processUnfinalizedBlockHeader(header?: Header): Promise<number | undefined> {
+    if (header) {
+      this.registerUnfinalizedBlock(header);
     }
 
     const forkedHeader = await this.hasForked();
@@ -105,6 +107,10 @@ export abstract class BaseUnfinalizedBlocksService<B> implements IUnfinalizedBlo
     }
 
     return;
+  }
+
+  async processUnfinalizedBlocks(block?: B): Promise<number | undefined> {
+    return this.processUnfinalizedBlockHeader(block ? this.blockToHeader(block) : undefined);
   }
 
   registerFinalizedBlock(header: Header): void {