+++ packages/node/src/indexer/worker/worker.unfinalizedBlocks.service.ts
@@ -2,11 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Injectable } from '@nestjs/common';
-import { IUnfinalizedBlocksService } from '@subql/node-core';
+import { Header, IUnfinalizedBlocksService } from '@subql/node-core';
 import { BlockContent } from '../types';
+import { substrateHeaderToHeader } from '../unfinalizedBlocks.service';
 
 export type HostUnfinalizedBlocks = {
-  unfinalizedBlocksProcess: (block: BlockContent) => Promise<number | null>;
+  unfinalizedBlocksProcess: (header: Header) => Promise<number | null>;
 };
 
 export const hostUnfinalizedBlocksKeys: (keyof HostUnfinalizedBlocks)[] = [
@@ -20,7 +21,13 @@ export class WorkerUnfinalizedBlocksService
   constructor(private host: HostUnfinalizedBlocks) {}
 
   async processUnfinalizedBlocks(block: BlockContent): Promise<number | null> {
-    return this.host.unfinalizedBlocksProcess(block);
+    return this.host.unfinalizedBlocksProcess(
+      substrateHeaderToHeader(block.block.block.header),
+    );
+  }
+
+  async processUnfinalizedBlockHeader(header: Header): Promise<number | null> {
+    return this.host.unfinalizedBlocksProcess(header);
   }
 
   // eslint-disable-next-line @typescript-eslint/promise-function-async