+++ packages/types/src/global.ts
@@ -12,5 +12,6 @@ declare global {
   const api: ApiAt;
   const logger: Pino.Logger;
   const store: Store;
+  const chainId: string;
   const createDynamicDatasource: DynamicDatasourceCreator;
 }